# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Инспектор ИИ
  version: 0.1.0-skeleton
  description: >
    Сверка ПД/РД/ИД по матрице 132 параметров. Асинхронная pull-модель:
    загрузка возвращает process_id, результат забирается по запросу.
    CONFIRMED_VIOLATION присваивает только инспектор.
    Имя продукта в интерфейсе — «Инспектор ИИ» (ТЗ п. 1.1).
    Кодовая линия репозитория — Контур.
servers:
  - url: /api/v1

# ТЗ п. 12: доступ только по токену. Роли операций перечислены в
# x-required-roles и продублированы в backend/src/kontur/presentation/rbac.py;
# расхождение ловит scripts/check_contracts.py.
security:
  - bearerAuth: []
paths:
...
  /system/capabilities:
    get:
      summary: Честный снимок слоёв извлечения
      description: >
        overall — живой путь (сейчас векторный текст). Отсутствующие OCR,
        таблицы и разбор чертежа перечислены явно: тишина не считается успехом.
        LLM/VLM — advisory и не валят комплект.
      operationId: getSystemCapabilities
      x-required-roles: [INSPECTOR, SUPERVISOR, ADMIN]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          description: Снимок возможностей
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemCapabilities" }
  /documents/upload:
    post:
      summary: Загрузка пакета документов (инкрементальная дозагрузка до финализации)
      operationId: uploadDocuments
      x-required-roles: [INSPECTOR, SUPERVISOR]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [object_id, files]
              properties:
                object_id: { type: string }
                process_id:
                  type: string
                  description: Указывается при дозагрузке в существующий процесс
                doc_stage:
                  type: string
                  enum: [PD, RD, ID]
                files:
                  type: array
                  # OAS 3.1 / JSON Schema 2020-12: contentMediaType вместо format:binary
                  items: { type: string, contentMediaType: application/octet-stream }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "202":
          description: Принято в обработку
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadAccepted" }
        "400": { $ref: "#/components/responses/Rejected" }
        "409":
          description: Протокол финализирован, дозагрузка запрещена
        "413":
          description: Превышен лимит файла (50 МБ) или пакета (200 МБ)
        "415":
          description: Неподдерживаемый формат файла (перечень — application.intake)
  /processes/{process_id}/status:
    get:
      summary: Статус обработки (целевой p95 ≤200 мс)
      operationId: getProcessStatus
      x-required-roles: [INSPECTOR, SUPERVISOR, ADMIN]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "404": { description: Процесс не найден }
  /processes/{process_id}/protocol:
    get:
      summary: Протокол проверки (структура по Приложению 2 ТЗ)
      operationId: getProtocol
      x-required-roles: [INSPECTOR, SUPERVISOR, ADMIN]
      parameters:
        - { $ref: "#/components/parameters/ProcessId" }
        - name: version
          in: query
          schema: { type: integer }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "./schemas/protocol.schema.json" }
        "404": { description: Процесс не найден или протокол ещё не собран }
  /processes/{process_id}/audit:
    get:
      summary: Журнал действий текущего процесса
      description: >
        События MemoryAudit этого процесса (REVIEW, START_VERIFICATION,
        COMPLETE_VERIFICATION, FINALIZE, UNFINALIZE, SELECT_REVISION).
        Это не таблица user_action_log: после рестарта журнал пуст (GAP-EDIT).
      operationId: getAuditLog
      x-required-roles: [INSPECTOR, SUPERVISOR, ADMIN]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditLog" }
        "404": { description: Процесс не найден }
  /findings/{finding_id}/review:
    post:
      summary: Решение инспектора по атомарной находке
      operationId: reviewFinding
      x-required-roles: [INSPECTOR, SUPERVISOR]
      parameters:
        - name: finding_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewDecision" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "./schemas/finding.schema.json" }
        "409": { description: Недопустимый переход состояния }
        "422": { description: Отсутствует reason_code при отклонении }
  /processes/{process_id}/verify:
    post:
      summary: Открыть очередь верификации
      description: >
        Человек переводит процесс READY → VERIFYING. Автомат и администратор
        очередь не открывают. Это не финализация протокола.
      operationId: startVerification
      x-required-roles: [INSPECTOR, SUPERVISOR]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inspector_id]
              properties:
                inspector_id: { type: string }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "404": { description: Процесс не найден }
        "409": { description: Не READY или вызывающий не инспектор }
  /processes/{process_id}/complete:
    post:
      summary: Закрыть очередь верификации
      description: >
        Человек переводит VERIFYING → COMPLETED после обработки всех CANDIDATE
        и SUSPICION. Не FINALIZED: финализация — отдельный POST /finalize.
      operationId: completeVerification
      x-required-roles: [INSPECTOR, SUPERVISOR]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inspector_id]
              properties:
                inspector_id: { type: string }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "404": { description: Процесс не найден }
        "409": { description: Есть необработанные кандидаты, не VERIFYING или не инспектор }
  /processes/{process_id}/finalize:
    post:
      summary: Финализация протокола
      description: >
        Разрешена только человеку и только после обработки всех CANDIDATE и
        SUSPICION. CLARIFICATION_REQUIRED — допустимый способ закрыть кандидата
        (п. 9.3). MISSING_EVIDENCE выводится отдельным перечнем.
      operationId: finalizeProtocol
      x-required-roles: [INSPECTOR, SUPERVISOR]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inspector_id]
              properties:
                inspector_id: { type: string }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "409": { description: Есть необработанные кандидаты или вызывающий не инспектор }
  /processes/{process_id}/unfinalize:
    post:
      summary: Отмена финализации супервизором
      description: >
        Только из PROTOCOL_FINALIZED / FINALIZED, только супервизор, только с
        причиной. Результат — COMPLETED, новая версия протокола и запись аудита
        (ADR-0005).
      operationId: unfinalizeProtocol
      x-required-roles: [SUPERVISOR]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inspector_id, reason]
              properties:
                inspector_id: { type: string }
                reason: { type: string, minLength: 1 }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "409": { description: Не FINALIZED или вызывающий не супервизор }
  /processes/{process_id}/revisions/{file_id}/select:
    post:
      summary: Инспектор назначает загруженный файл эталоном редакции
      description: >
        Человек подтверждает, что файл — последняя применимая утверждённая
        редакция для сравнения. Автомат и администратор не вызывают операцию.
        Штамп «не утв.» нельзя перекрыть. «ГИП» и пустая графа «Утвердил»
        сами по себе эталоном не становятся. После выбора матрица
        пересчитывается; CONFIRMED_VIOLATION не пишется. Только из READY.
        Это не закрытие гейта J и не замена штампа.
      operationId: selectRevision
      x-required-roles: [INSPECTOR, SUPERVISOR]
      parameters:
        - { $ref: "#/components/parameters/ProcessId" }
        - { $ref: "#/components/parameters/FileId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inspector_id, comment]
              properties:
                inspector_id: { type: string }
                comment: { type: string, minLength: 1 }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatus" }
        "404": { description: Процесс или файл не найден }
        "409": { description: Не READY, финализирован, штамп NOT_APPROVED или нет содержимого }
  /inspection/{process_id}:
    post:
      summary: Передача результатов во внешнюю ИС (целевая — ИАИС «РиН»)
      description: >
        Разрешена только при PROTOCOL_FINALIZED. Передаются только
        подтверждённые инспектором записи с версиями протокола, матрицы,
        модели и реестром входных файлов. До 3 повторов: 1, 5, 15 минут.
      operationId: syncInspection
      x-required-roles: [SUPERVISOR, ADMIN]
      parameters: [{ $ref: "#/components/parameters/ProcessId" }]
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "202":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncState" }
        "409": { description: Протокол не финализирован }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT проверяется только с одним настроенным асимметричным алгоритмом:
        RS256 с RSA-ключом либо ES256 с ключом P-256. Валидируются подпись,
        iss, aud, exp, nbf, непустой sub и массив roles с ролями INSPECTOR /
        SUPERVISOR / ADMIN. Операции, привязанные к объекту, дополнительно
        требуют непустой object_id в токене, совпадающий с объектом запроса;
        несовпадение запрещает доступ. GET /healthz публичен. GET
        /system/capabilities требует JWT, но не требует object_id.
        Администратор не подтверждает нарушения и не финализирует протокол
        (ADR-0001), отмена финализации — только супервизор (п. 9.3).
...
  responses:
    Unauthorized:
      description: Аутентификация не пройдена — JWT отсутствует, недействителен или не прошёл проверку обязательных claims
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: JWT действителен, но роль, object_id или требование инспектора не допускают операцию
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    UploadAccepted:
      type: object
      required: [process_id, accepted, rejected]
      properties:
        process_id: { type: string }
        accepted:
          type: array
          items:
            type: object
            required: [file_id, file_hash]
            properties:
              file_id: { type: string }
              file_hash: { type: string, pattern: "^[a-f0-9]{64}$" }
        rejected:
          type: array
          items: { $ref: "#/components/schemas/RejectionReason" }
...
    SystemCapabilities:
      type: object
      required: [overall, kit_blocked, engines, health]
      properties:
        overall:
          type: string
          enum: [AVAILABLE, DEGRADED, UNAVAILABLE]
          description: Статус живого пути извлечения, не «все слои зелёные».
        kit_blocked: { type: boolean }
        note: { type: string }
        engines:
          type: array
          items:
            type: object
            required: [name, status, affects_verdict]
            properties:
              name: { type: string }
              status:
                type: string
                enum: [AVAILABLE, DEGRADED, UNAVAILABLE]
              affects_verdict: { type: boolean }
        health:
          type: object
          required: [healthy, degraded, failed, skipped]
          properties:
            healthy:
              type: array
              items: { type: string }
            degraded:
              type: array
              items: { type: string }
            failed:
              type: array
              items: { type: string }
            skipped:
              type: array
              items: { type: string }
...